chore(): move swapSingleUser to /utils
